requirements.txt
discord.py>=2.3.0
google-generativeai>=0.5.0
chromadb>=0.4.0
sentence-transformers>=2.7.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
